requirements.txt
cyglfw3
pyopengl==3.1.1a1